Extact test cases out
